--- packages/code-du-travail-frontend/src/modules/integration/index.ts
@@ -1,13 +1,0 @@
-export { IntegrationList } from "./IntegrationList";
-export type { IntegrationListProps } from "./IntegrationList";
-
-export { IntegrationPageContent } from "./IntegrationPageContent";
-
-export { IntegrationDetailContent } from "./IntegrationDetailContent";
-export type {
-  IntegrationDetailContentProps,
-  SelectItem,
-} from "./IntegrationDetailContent";
-
-export { IntegrationDetailPageContent } from "./IntegrationDetailPageContent";
-export type { IntegrationDetailPageContentProps } from "./IntegrationDetailPageContent";
